Refactor feature:clock packaging
