+++ feature/clock/src/main/java/com/mmb/clock/TimerUtils.kt
@@ -1,6 +1,7 @@
 package com.mmb.clock
 
-import com.mmb.clock.TimerState.Companion.SECONDS_PER_MINUTE
+import com.mmb.clock.entity.TimerState
+import com.mmb.clock.entity.TimerState.Companion.SECONDS_PER_MINUTE
 
 fun Long.convertToTimerState(): TimerState {
     val min: Long = this / SECONDS_PER_MINUTE
